Финальная версия проекта

+++ functions.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 
 function coding {
+    clear
+
+    counter=0
     IFS=:
     red_close="\033[0m"
+    echo "Пишите код, нажимая различные клавиши, а когда закончите, пафосно нажмите <enter>!"
 
-    while true
+    while [[ true ]]
     do
         for line in $file
         do
@@ -15,6 +19,7 @@ function coding {
                 if [[ $k ]]
                 then
                     printf "$color_open$word$red_close "
+                    counter=$(($counter + 1))
                     continue
                 else
                     printf "\n"
@@ -36,68 +41,87 @@ function style_select {
     printf 'Добрый день! Вас приветствует программа для взлома Пентагона \033[92mDanyaHacker.com\033[0m \n'
     PS3='Выберите стиль оформления: '
 
-    select color in Red_blood Green_hacker Orange_mango Blue_sea Purple_haze Blue_sky White_standart
-    do
-        case $color in
-        Red_blood)
-            color_open='\033[91m'
-            ;;
-        Green_hacker) 
-            color_open='\033[92m'
-            ;;
-        Orange_mango)
-            color_open='\033[93m'
-            ;;
-        Blue_sea)
-            color_open='\033[94m'
-            ;;
-        Purple_haze)
-            color_open='\033[95m'
-            ;;
-        Blue_sky)
-            color_open='\033[96m'
-            ;;
-        White_standart)
-            color_open='\033[97m'
-            ;;
-        quit)
+    while [[ true ]]; do
+        select color in Red_blood Green_hacker Orange_mango Blue_sea Purple_haze Blue_sky White_standart
+        do
+            case $color in
+            Red_blood)
+                color_open='\033[91m'
+                language_select
+                break
+                ;;
+            Green_hacker) 
+                color_open='\033[92m'
+                language_select
+                break
+                ;;
+            Orange_mango)
+                color_open='\033[93m'
+                language_select
+                break
+                ;;
+            Blue_sea)
+                color_open='\033[94m'
+                language_select
+                break
+                ;;
+            Purple_haze)
+                color_open='\033[95m'
+                language_select
+                break
+                ;;
+            Blue_sky)
+                color_open='\033[96m'
+                language_select
+                break
+                ;;
+            White_standart)
+                color_open='\033[97m'
+                language_select
+                break
+                ;;
+            quit)
+                break
+                ;;
+            *) 
+                printf "\033[91mНедопустимая опция $REPLY\033[0m \n"
+                ;;
+            esac
             break
-            ;;
-        *) 
-            printf "\033[91mНедопустимая опция $REPLY\033[0m"
-            ;;
-        esac
-        break
+        done
     done
-    language_select
 }
 
 function language_select {
     echo 'Замечательно, со стилем мы определились, теперь выберите язык программирования, на котором хотите взломать Пентагон'
     PS3="Мой выбор: "
 
-    select language in Python C# 
-    do
-        case $language in
-        Python)
-            file=$(cat ./scripts/Python.txt)
-            ;;
-        C#)
-            file=$(cat ./scripts/C#.txt)
-            ;;
-        quit)
+    while [[ true ]]; do
+        select language in Python C# 
+        do
+            case $language in
+            Python)
+                file=$(cat ./scripts/Python.txt)
+                echo "Супер, пора приступать!"
+                loading_icon 8 "Начинаем взлом!!!"
+                coding
+                ;;
+            C#)
+                file=$(cat ./scripts/C#.txt)
+                echo "Супер, пора приступать!"
+                loading_icon 8 "Начинаем взлом!!!"
+                coding
+                ;;
+            quit)
+                break
+                ;;
+            *) 
+                printf "\033[91mНедопустимая опция $REPLY\033[0m \n"
+                ;;
+            esac
             break
-            ;;
-        *) 
-            printf "\033[91mНедопустимая опция $REPLY\033[0m"
-            ;;
-        esac
-        break
+        done
     done
-
-    echo "Супер, пора приступать!"
-    loading_icon 8 "Начинаем взлом!!!"
-    coding
 }
 
 function loading_icon {
@@ -124,17 +148,23 @@ function loading_icon {
 
 function ending {
     loading_icon 8 "Взлом Пентагона!!!"
-    # number=$(( $RANDOM % 2))
-    number=1
-    if [[ $number -eq 1 ]]
-    then
-        echo "Пентагон успешно взломан, поздравляю!!!!"
-        mining_choise
-    else 
-        echo "Увы, Пентагон не взломан!"
-        echo "Вас обнаружили, все данные будут стёрты через 5 секунд!!!!"
-        sleep 5
-        poweroff
+    clear
+
+    if [[ $counter -lt 50 ]]; then
+        echo "Увы, этого было неостаточно, Пентагон не взломан! :("
+        exit
+    else
+        number=$(( $RANDOM % 2))
+        if [[ $number -eq 1 ]]
+        then
+            echo "Пентагон успешно взломан, поздравляю!!!!"
+            mining_choise
+        else 
+            echo "Увы, Пентагон не взломан!"
+            echo "Вас обнаружили, все данные будут стёрты через 5 секунд!!!!"
+            sleep 5
+            poweroff
+        fi
     fi
 }
 
@@ -147,14 +177,19 @@ function mining {
     while true
     do
         read -s -n 1 k <&1
-        if [ $k != 'q' ]
+        if [[ $k != 'q' ]]
         then
             bitcoin=$(( $RANDOM % 5 ))
             printf "$color_open + $bitcoin Bitcoin $red_close \n"
             total=$(( $total + $bitcoin ))
             continue
+        elif [[ $k == ' ' ]]; then
+            printf "Вы намайнили $color_open$total Биткоинов!!!$red_close \n"
+            printf "Это $color_open$(($total * 66773 )) долларов$red_close, или $color_open$(($total * 6069052 )) рублей!$red_close \n"
+            break
         else
-            printf " Вы намайнили $total Биткоинов \n"
+            printf "Вы намайнили $color_open$total Биткоинов!!!$red_close \n"
+            printf "Это $color_open$(($total * 66773 )) долларов$red_close, или $color_open$(($total * 6069052 )) рублей!$red_close \n"
             break
         fi
     done
@@ -166,14 +201,20 @@ function mining {
 function mining_choise {
     read -p "Хотите взломать счёт одного из сотрудников Пентагона и украсть его Биткоины? [y/n]: " choice
 
-    if [[ $choice == 'y' ]]
-    then
-        echo "Отлично!"
-        mining
-    else
-        echo "Правильно, лучше не рисковать, скрываемся!!!!"
-        exit
-    fi
+    while [[ true ]]; do    
+        if [[ $choice == 'y' ]]
+        then
+            echo "Отлично!"
+            mining
+            break
+        elif [[ $choice == 'n' ]]
+        then
+            echo "Правильно, лучше не рисковать, скрываемся!!!!"
+            exit
+        else
+            read -p "Нет такой команды, введите <y> если вы согласны, и <n> если нет: " choice
+        fi
+    done
 }
 
 style_select
